requirements.txt
streamlit
langchain_community
langchain_groq
langchain_core
chromadb
pandas
validators
python-dotenv